Conflicts:
	basebmp/prj/d.lst
	basebmp/test/basictest.cxx
	basebmp/test/makefile.mk
	basegfx/inc/basegfx/basegfxdllapi.h
	basegfx/inc/basegfx/tools/debugplotter.hxx
	basegfx/inc/basegfx/tuple/b2ituple.hxx
	basegfx/prj/d.lst
	basegfx/source/numeric/makefile.mk
	basegfx/source/polygon/makefile.mk
	basegfx/source/range/makefile.mk
	basegfx/source/raster/makefile.mk
	basegfx/source/tuple/makefile.mk
	basegfx/source/vector/makefile.mk
	basegfx/test/basegfx1d.cxx
	basegfx/test/makefile.mk
	basegfx/util/makefile.mk
	canvas/Library_canvasfactory.mk
	canvas/Module_canvas.mk
	canvas/prj/build.lst
	canvas/prj/d.lst
	canvas/source/cairo/cairo_canvashelper_texturefill.cxx
	canvas/source/cairo/makefile.mk
	canvas/source/tools/makefile.mk
	comphelper/qa/string/makefile.mk
	cppcanvas/Module_cppcanvas.mk
	cppcanvas/inc/cppcanvas/cppcanvasdllapi.h
	cppcanvas/prj/build.lst
	cppcanvas/prj/d.lst
	cppcanvas/source/mtfrenderer/makefile.mk
	cppcanvas/util/makefile.mk
	i18npool/source/search/makefile.mk
	regexp/Library_regexp.mk
	regexp/prj/d.lst
	sax/CppunitTest_sax.mk
	sax/Library_sax.mk
	sax/prj/d.lst
	sax/qa/cppunit/test_converter.cxx
	sax/source/expatwrap/attrlistimpl.hxx
	sax/util/makefile.mk
	svtools/Library_svt.mk
	tools/Executable_sspretty.mk
	ucbhelper/prj/d.lst
	ucbhelper/source/provider/configureucb.cxx
	ucbhelper/source/provider/provconf.cxx
	ucbhelper/util/makefile.mk
	unotools/Library_utl.mk
	unotools/Module_unotools.mk
	unotools/Package_inc.mk
	unotools/prj/build.lst
	vcl/Library_desktop_detector.mk
	vcl/Library_vcl.mk
	vcl/Library_vclplug_gtk.mk
	vcl/aqua/source/gdi/salprn.cxx
	vcl/inc/aqua/saldata.hxx
	vcl/unx/generic/gdi/salgdi3.cxx

+++ vcl/unx/gtk/a11y/atkwindow.cxx
@@ -1,3 +1,4 @@
+/* -*- Mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*- */
 /*************************************************************************
  *
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
@@ -295,13 +296,13 @@ ooo_window_wrapper_get_type (void)
 
         static const GTypeInfo typeInfo =
         {
-            type_query.class_size,
+            static_cast<guint16>(type_query.class_size),
             (GBaseInitFunc) NULL,
             (GBaseFinalizeFunc) NULL,
             (GClassInitFunc) ooo_window_wrapper_class_init,
             (GClassFinalizeFunc) NULL,
             NULL,
-            type_query.instance_size,
+            static_cast<guint16>(type_query.instance_size),
             0,
             (GInstanceInitFunc) NULL,
             NULL
@@ -329,3 +330,4 @@ void restore_gail_window_vtable (void)
     atk_class->initialize = window_real_initialize;
 }
 
+/* vim:set shiftwidth=4 softtabstop=4 expandtab: */
